Perform major refactor to move covid-specific logic to covid flavor
