+++ rdt-open-reader/src/main/java/io/ona/rdt/interactor/PatientProfileActivityInteractor.java
@@ -1,24 +1,9 @@
 package io.ona.rdt.interactor;
 
-import org.json.JSONObject;
-
-import io.ona.rdt.callback.OnFormSavedCallback;
+import io.ona.rdt.util.FormSaver;
 
 /**
  * Created by Vincent Karuri on 17/02/2020
  */
-public class PatientProfileActivityInteractor {
-
-    private PatientRegisterFragmentInteractor patientRegisterFragmentInteractor;
-
-    public void saveForm(JSONObject jsonForm, OnFormSavedCallback callback) {
-        getPatientRegisterFragmentInteractor().saveForm(jsonForm, callback);
-    }
-
-    public PatientRegisterFragmentInteractor getPatientRegisterFragmentInteractor() {
-        if (patientRegisterFragmentInteractor == null) {
-            patientRegisterFragmentInteractor = new PatientRegisterFragmentInteractor();
-        }
-        return patientRegisterFragmentInteractor;
-    }
+public class PatientProfileActivityInteractor extends FormSaver {
 }
