+++ rdt-open-reader/src/main/java/io/ona/rdt/interactor/RDTJsonFormFragmentInteractor.java
@@ -1,0 +1,9 @@
+package io.ona.rdt.interactor;
+
+import io.ona.rdt.util.FormSaver;
+
+/**
+ * Created by Vincent Karuri on 13/07/2020
+ */
+public class RDTJsonFormFragmentInteractor extends FormSaver {
+}
